requirements.txt
# Core
sqlalchemy>=2.0.0
typer[all]>=0.9.0
rich>=13.0.0

# API clients
anthropic>=0.40.0
httpx>=0.27.0
requests>=2.31.0

# Gmail OAuth
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6

# CV generation
python-docx>=1.1.0
docx2pdf>=0.1.8

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Auth
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.0
bcrypt>=4.0.0

# Utilities
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
